+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 ### Added
+- Add locale_path to settings, to help facilitate translation when required
 - Removed celery settings, as it's hardly used.
 - Add api versioning support via Accept headers (#52)
 - Add `favicon.ico`, `robots.txt`, `humans.txt`